requirements.txt
altgraph==0.17.4
certifi==2025.8.3
charset-normalizer==3.4.3
et_xmlfile==2.0.0
idna==3.10
numpy==2.3.3
openpyxl==3.1.5
packaging==25.0
pandas==2.3.2
pyinstaller==6.16.0
pyinstaller-hooks-contrib==2025.9
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
setuptools==80.9.0
six==1.17.0
tzdata==2025.2
urllib3==2.5.0
